Create a test split with labels to enable linear probes
