+++ src/squirrel_datasets_core/datasets/AutoML/driver.py
@@ -5,10 +5,13 @@ import zipfile
 from typing import TYPE_CHECKING, Any, Dict, List
 
 import numpy as np
+import pandas as pd
 import requests
 
 from squirrel.driver import IterDriver
 from squirrel.iterstream import IterableSource
+from squirrel_datasets_core.datasets.utils import proportionate_sample_df
+
 
 if TYPE_CHECKING:
     from squirrel.iterstream import Composable
@@ -29,7 +32,7 @@ class AutoML(IterDriver):
         self.dataset_name = dataset_name
         self.zipfile = zipfile.ZipFile(io.BytesIO(requests.get(URLS[self.dataset_name]).content))
 
-    def _get_train_split(self) -> List[Dict[str, Any]]:
+    def _get_train_test_split(self, split: str = "train") -> List[Dict[str, Any]]:
         """Create train split"""
         records = []
         for feat, lbl in zip(
@@ -37,22 +40,32 @@ class AutoML(IterDriver):
             self.zipfile.read(f"{self.dataset_name}_train.solution").decode("utf-8").split("\n"),
         ):
             try:
-                records.append(
-                    {
-                        "features": [float(x) for x in feat.strip().split(" ")],
-                        "class": int(np.argwhere(np.asarray([int(x) for x in lbl.strip().split(" ")]) == 1)),
-                    }
-                )
+                record = {
+                    f"features_{idx:03d}": val for idx, val in enumerate([float(x) for x in feat.strip().split(" ")])
+                }
+                record["class"] = int(np.argwhere(np.asarray([int(x) for x in lbl.strip().split(" ")]) == 1))
+                records.append(record)
             except ValueError:
                 pass
-        return records
 
-    def _get_test_or_valid_split(self, split: str = "train") -> List[Dict[str, Any]]:
+        df = pd.DataFrame.from_records(records)
+        train_df, test_df = proportionate_sample_df(df, "class", 0.2, seed=42)
+        if split == "train":
+            return train_df.to_dict(orient="records")
+        else:
+            return test_df.to_dict(orient="records")
+
+    def _get_test_or_valid_split(self, split: str = "orig_train") -> List[Dict[str, Any]]:
         """Create test or validation split"""
+        split = split.split("_")[-1]
         records = []
         for feat in self.zipfile.read(f"{self.dataset_name}_{split}.data").decode("utf-8").split("\n"):
             try:
-                records.append({"features": [float(x) for x in feat.strip().split(" ")], "class": None})
+                record = {
+                    f"features_{idx:03d}": val for idx, val in enumerate([float(x) for x in feat.strip().split(" ")])
+                }
+                record["class"] = None
+                records.append(record)
             except ValueError:
                 pass
         return records
@@ -66,8 +79,10 @@ class AutoML(IterDriver):
             shuffle_item_buffer (int): the size of the buffer used to shuffle samples after being fetched.
                 Please note the memory footprint of samples
         """
-        assert split in ["train", "test"]
+        assert split in ["train", "test", "orig_test", "orig_valid"]
         if split == "train":
-            return IterableSource(self._get_train_split()).shuffle(size=shuffle_item_buffer)
-        else:
+            return IterableSource(self._get_train_test_split(split="train")).shuffle(size=shuffle_item_buffer)
+        elif split == "test":
+            return IterableSource(self._get_train_test_split(split="test")).shuffle(size=shuffle_item_buffer)
+        elif split in ["orig_train", "orig_valid"]:
             return IterableSource(self._get_test_or_valid_split(split)).shuffle(size=shuffle_item_buffer)
